Tim11/requirements.txt
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
django-jet==1.0.8
django-location-field==2.1.0
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.8.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
